requirements.txt
Django==1.10.3
django-haystack==2.5.1
django-widget-tweaks==1.4.1
django-recaptcha==1.0.5
django-cleanup==0.4.2
django-material
python-magic==0.4.12
django-tinymce
django-extensions
django-google-maps
pysolr
Pillow
psycopg2

